fix(ppt-web): #486 — centralize authenticated fetch in api-client, useCallback filter handlers, use usePinnedAnnouncements
